Beginning of Onion Architecture
