+++ Chirp.sln
@@ -11,6 +11,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{3586FA51-7
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chrip.razor.test", "test\Chrip.razor.test\Chrip.razor.test.csproj", "{7CF8BBE8-00A9-4FAE-A261-90C106E4DE0D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Core", "src\Chirp.Core\Chirp.Core.csproj", "{A395319B-87A5-469D-9011-62234ADF8FF1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Infrastructure", "src\Chirp.Infrastructure\Chirp.Infrastructure.csproj", "{01012D57-05E3-42A7-A11B-7BAB3123DCBB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Web", "src\Chirp.Web\Chirp.Web.csproj", "{1635E6C2-4CC7-4C66-916E-85DA6E28925C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,9 +34,24 @@ Global
 		{7CF8BBE8-00A9-4FAE-A261-90C106E4DE0D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7CF8BBE8-00A9-4FAE-A261-90C106E4DE0D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7CF8BBE8-00A9-4FAE-A261-90C106E4DE0D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A395319B-87A5-469D-9011-62234ADF8FF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A395319B-87A5-469D-9011-62234ADF8FF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A395319B-87A5-469D-9011-62234ADF8FF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A395319B-87A5-469D-9011-62234ADF8FF1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{01012D57-05E3-42A7-A11B-7BAB3123DCBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{01012D57-05E3-42A7-A11B-7BAB3123DCBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{01012D57-05E3-42A7-A11B-7BAB3123DCBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{01012D57-05E3-42A7-A11B-7BAB3123DCBB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1635E6C2-4CC7-4C66-916E-85DA6E28925C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1635E6C2-4CC7-4C66-916E-85DA6E28925C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1635E6C2-4CC7-4C66-916E-85DA6E28925C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1635E6C2-4CC7-4C66-916E-85DA6E28925C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{BC54E309-EA36-47E0-852E-248B47C48BB3} = {3F64737C-074F-4229-943E-41822EF6E0F5}
 		{7CF8BBE8-00A9-4FAE-A261-90C106E4DE0D} = {3586FA51-721F-498F-A9AA-B2E543B05A6B}
+		{A395319B-87A5-469D-9011-62234ADF8FF1} = {3F64737C-074F-4229-943E-41822EF6E0F5}
+		{01012D57-05E3-42A7-A11B-7BAB3123DCBB} = {3F64737C-074F-4229-943E-41822EF6E0F5}
+		{1635E6C2-4CC7-4C66-916E-85DA6E28925C} = {3F64737C-074F-4229-943E-41822EF6E0F5}
 	EndGlobalSection
 EndGlobal
