+++ src/Chirp.Web/Pages/Shared/_Layout.cshtml
@@ -1,0 +1,22 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>@ViewData["Title"]</title>
+    <link href="/css/style.css" rel="stylesheet" type="text/css" />
+    <link rel="icon" type="image/png" href="/favicon/favicon.ico">
+</head>
+<body>
+<div class=page>
+    <h1><img src="/images/icon1.png" alt="Icon1"/>Chirp!</h1>
+
+    <div class=body>
+            @RenderBody()
+    </div>
+    <div class=footer>
+        Chirp &mdash; An ASP.NET Application
+    </div>
+</div>
+
+</body>
+@await RenderSectionAsync("Scripts", required: false)
+</html>
