+++ src/Chirp.Web/wwwroot/css/style.css
@@ -1,0 +1,289 @@
+body {
+  background: #eccaca;
+  font-family: 'Trebuchet MS', sans-serif;
+  font-size: 14px;
+}
+
+a {
+  color: #26776F;
+}
+
+a:hover {
+  color: #333;
+}
+
+input[type="email"],
+input[type="text"],
+input[type="tel"],
+input[type="password"] {
+    background: white;
+    border: 1px solid #e6c0bf;
+    padding: 2px;
+    font-family: 'Trebuchet MS', sans-serif;
+    font-size: 14px;
+    border-radius: 2px;
+    -moz-border-radius: 2px;
+    -webkit-border-radius: 2px;
+    color: #571010;
+}
+
+input[type="submit"] {
+  background: #571010;
+  border: 1px solid #3b0707;
+  padding: 1px 3px;
+  font-family: 'Trebuchet MS', sans-serif;
+  font-size: 14px;
+  font-weight: bold;
+  border-radius: 2px;
+  -moz-border-radius: 2px;
+  -webkit-border-radius: 2px;
+  color: white;
+}
+
+div.page {
+  background: white;
+  border: 1px solid #cc6e6e;
+  width: 700px;
+  margin: 30px auto;
+}
+
+div.page > h1 {
+  background: #cc6e6e;
+  margin: 0;
+  padding: 10px 14px;
+  color: white;
+  letter-spacing: 1px;
+  text-shadow: 0 0 3px #772424;
+  font-weight: normal;
+  font-size: 80px;
+}
+
+div.page h1 img {
+  height: 1em;
+}
+
+div.page div.navigation {
+  background: #DEE9E8;
+  padding: 4px 10px;
+  border-top: 1px solid #ccc;
+  border-bottom: 1px solid #eee;
+  color: #888;
+  font-size: 12px;
+  letter-spacing: 0.5px;
+}
+
+div.page div.navigation a {
+  color: #444;
+  font-weight: bold;
+}
+
+div.page div.navigation form {
+    display: contents;
+}
+
+div.page div.navigation form button {
+    padding: unset;
+    background: unset;
+    color: #444;
+    font-weight: bold;
+    display: inline-block;
+    font-size: unset;
+    cursor: pointer;
+    text-decoration: underline;
+    font-family: unset;
+    letter-spacing: unset;
+}
+
+div.page h2 {
+  margin: 0 0 15px 0;
+  color: #571010;
+  text-shadow: 0 1px 2px #ccc;
+}
+
+div.page div.body {
+  padding: 10px;
+}
+
+div.page div.footer {
+  background: #eee;
+  color: #888;
+  padding: 5px 10px;
+  font-size: 12px;
+}
+
+div.page div.followstatus {
+  border: 1px solid #ccc;
+  background: #E3EBEA;
+  border-radius: 2px;
+  -moz-border-radius: 2px;
+  -webkit-border-radius: 2px;
+  padding: 3px;
+  font-size: 13px;
+}
+
+div.page ul.cheeps {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+div.page ul.cheeps li {
+  margin: 10px 0;
+  padding: 5px;
+  background: #F0FAF9;
+  border: 1px solid #DBF3F1;
+  border-radius: 2px;
+  -moz-border-radius: 5px;
+  -webkit-border-radius: 5px;
+  min-height: 48px;
+}
+
+div.page ul.cheeps p {
+  margin: 0;
+}
+
+div.page ul.cheeps li img {
+  float: left;
+  padding: 0 10px 0 0;
+}
+
+div.page ul.cheeps li small {
+  font-size: 0.9em;
+  color: #888;
+}
+
+div.page div.cheepbox {
+  margin: 10px 0;
+  padding: 5px;
+  background: #F0FAF9;
+  border: 1px solid #e29494;
+  border-radius: 2px;
+  -moz-border-radius: 5px;
+  -webkit-border-radius: 5px;
+}
+
+div.page div.cheepbox h3 {
+  margin: 0;
+  font-size: 1em;
+  color: #7e2c2c;
+}
+
+div.page div.cheepbox p {
+  margin: 0;
+}
+
+div.page div.cheepbox input[type="text"] {
+  width: 585px;
+}
+
+div.page div.cheepbox input[type="submit"] {
+  width: 70px;
+  margin-left: 5px;
+}
+
+div.message {
+  list-style: none;
+  margin: 10px 10px 0 10px;
+
+  background: #f3b9b9;
+  border: 1px solid #ce8181;
+  border-radius: 2px;
+  -moz-border-radius: 2px;
+  -webkit-border-radius: 2px;
+  padding: 4px;
+  font-size: 13px;
+}
+
+ul.message li {
+  background: #f3b9b9;
+  border: 1px solid #ce8181;
+  border-radius: 2px;
+  -moz-border-radius: 2px;
+  -webkit-border-radius: 2px;
+  padding: 4px;
+  font-size: 13px;
+}
+
+div.error {
+  margin: 10px 0;
+  background: #FAE4E4;
+  border: 1px solid #DD6F6F;
+  border-radius: 2px;
+  -moz-border-radius: 2px;
+  -webkit-border-radius: 2px;
+  padding: 4px;
+  font-size: 13px;
+}
+
+div.topics {
+  width: 180px;
+  height: 250px;
+  background: #ffffff;
+  border: 1px solid #e29494;
+  border-radius: 2px;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  margin-left: 460px;
+  padding: 10px;
+}
+
+ul.topic {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+li.topic {
+  margin: 10px 0;
+  padding: 5px;
+  background: #F0FAF9;
+  border: 1px solid #DBF3F1;
+  border-radius: 2px;
+  -moz-border-radius: 5px;
+  -webkit-border-radius: 5px;
+  min-height: 20px;
+}
+
+form * {
+    margin: 2px 0;
+}
+
+button {
+    background-color: #cc6e6e;
+    border: none;
+    color: white;
+    padding: 8px 16px;
+    text-align: center;
+    text-decoration: none;
+    cursor: pointer;
+}
+
+button:hover {
+    background-color: #8b4949;
+}
+
+ul.nav {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    overflow: hidden;
+    background-color: #cc6e6e;
+}
+
+ul.nav li {
+    float: left;
+}
+
+ul.nav li a {
+    display: block;
+    color: white;
+    text-align: center;
+    padding: 8px 8px;
+    text-decoration: none;
+}
+
+ul.nav li a:hover {
+    background-color: #8b4949;
+}
