--- src/Chirp.razor/Pages/Public.cshtml
@@ -1,32 +1,0 @@
-﻿@page "/"
-@model Chirp.Pages.PublicModel
-@{
-    ViewData["Title"] = "Chirp!";
-    Layout = "Shared/_Layout";
-}
-
-<div>
-    <h2> Public Timeline </h2>
-
-    @if (Model.Cheeps.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            @foreach (var cheep in Model.Cheeps)
-            {
-                <li>
-                    <p>
-                        <strong>
-                            <a href="/@cheep.Author">@cheep.Author</a>
-                        </strong>
-                        @cheep.Text
-                        <small>&mdash; @cheep.TimeStamp</small>
-                    </p>
-                </li>
-            }
-        </ul>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
-</div>
