--- src/Chirp.razor/Pages/UserTimeline.cshtml
@@ -1,33 +1,0 @@
-﻿@page "/{author}"
-@model Chirp.Pages.UserTimelineModel
-@{
-    ViewData["Title"] = "Chirp!";
-    Layout = "Shared/_Layout";
-    var routeName = HttpContext.GetRouteValue("author");
-}
-
-<div>
-    <h2> @routeName's Timeline </h2>
-
-    @if (Model.Cheeps.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            @foreach (var cheep in Model.Cheeps)
-            {
-                <li>
-                    <p>
-                        <strong>
-                            <a href="/@cheep.Author">@cheep.Author</a>
-                        </strong>
-                        @cheep.Text
-                        <small>&mdash; @cheep.TimeStamp</small>
-                    </p>
-                </li>
-            }
-        </ul>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
-</div>
